Fix chipram synthesis to no longer rely on ISE magic to get the right behavior.
The main fix was to make sure that the ramaddress signal was calculated via combinatorial logic and not as a register.  This allows it to be fed as input into the block ram during the clock cycle where the new value is calculated so that the correct output value is available on the following clock cycle.