Make colourram synchronous.
It turns out the viciv state machine is basically already taking into account the extra 2 clock delay, although it should be possible to fix so it's not really needed.  But for now this eliminates a massive RAM that was eating up a lot of resources.   For my 15" MBP the map time went from about an hour and a half to more like 38 minutes.