Finally get shadowram, romram, and kickstart ram wokring as synchronous (block) rams.
This involved some pretty major refactoring of the CPU's address generation logic to be in a separate combinatorial block so that it can provide the "next" address (primarily reads) during the current state clock phase.   This is essentially what ISE has been doing internally in order to make the asynchronously-coded read ports on the RAMs work with the synchronous block RAMs.  Rather than have that behavior get inferred, it's now done explicitly.

For the kickstart RAM I had to plumb through a separate external address bus that provides the read address a clock earlier than the fastio port normally expects.  I did it this way so I wouldn't change any other fastio behavior yet.

Yes my code is still a bit of a mess, and it's still causing Vivado to infer a few latches I don't want, but those are easily fixable.

It appears the next RAM that has to get fixed are the sdcardio RAMs.   This will also possibly require a fastio address bus that has the address a cycle earlier, so I may just repurpose what I did for the kickstart RAM and always provide a parallel "fastio_sync" bus or something that I can have different modules adopt one at a time.