Get sdcardio sector buffer using an explicit synchronous ram.
This required plumbing an explicit "fast" fastio_addr signal out of the CPU core and into the rest of the machine so that things that can respond faster can pick it up a cycle earlier.  This probably ought to get merged with the faster kickstart_addr signal at some point.   Eventually this faster address output ought to let some of the extra wait states on fastio devices go away entirely assuming timing constraints can be met.