Success!  Vivado version now boots all the way to C65 BASIC prompt at least.
The final issue was dealing with the fact that Vivado appears to treat internal 'Z' values as 0 when not driven instead of 1, so the keyboard and IRQ lines were all acting weird.

Bouldermark at least starts up but eventually goes a little wonky and exits.   Need to see if there are any other obvious things that need to be fixed in regards to interrupts.