Some simplification of address handling.
Try to unify upper and lower map enables, and have there be only a single address adder and not duplicated paths.

Don't mess with bits 12-15 where not needed.  In many cases they already had the right value.

For cases where we know we are mapping to the external port, just remember that and just do that remap once.

Rip out some of the colourram at 7f4xxx stuff since nothing is depending on it at the moment since that helps a bit too.