Simplify (yes, really) DMA job register loading.
It actually requires fewer resources to load the registers as we read, rather than have a datapath that has to string through all of them.

Doing it based on the current count (like a small state machine) means that most of the cases (and bits) can share the same LUTs for the load signals.