First pass at new DMAgic implementation.
This gets the basics working of CPU driven read and write cycles.  The implementation is horrible but proves it can work.

I can now do arbitrary flat 20-bit addressing with or without auto-incrementing reads and writes, which should let me get rid of the 32-bit load/store stuff in the CPU core.

It also lays the groundwork for moving the rest of the DMAgic implementation to the new external unit where it should be.