Implement a new and simpler timing emulation scheme that doesn't use the instruction lookup tables.
Also, added a special 25Mhz speed mode to toggle phi_en to stress test the signaling.  Found a few bugs in DMAgic implementation where I needed to be using the CPU ack signal to ignore cycles where the CPU is held idle.   I didn't notice before because DMAgic is barely used (if at all) once the C65 boots and switches to 3.5Mhz.

Eventually I want to move the clock speed control stuff out of the CPU core logic, so this was a step toward being able to do that.