Fix the instability bug with interrupts.
The interrupt logic was setting the intg flag early when the bus wasn't ready and we weren't moving to the next clock cycle yet.  This was causing the PC increment to be disabled when it shouldn't have been, which meant that the address pushed on the stack when an interrupt happened was wrong.