Add a wait state to dmagic I/O registers to help improve timing.
This was really tricky for a few reasons.  The main one being that I had to fix the DMAgic io_ready signal to only drive 1 once a transfer had been done, rather than have it default to one.   This only really worked by before because DMAgic had zero wait states and was basically seeing the next address bus address the clock before the ready signal needed to be updated.  Now that DMAgic doesn't get that early peek, I can't have the ready signal default to one.

So, it took a little bit of extra logic to only drive the 1 once we do a "safe" register access or the PIO DMA operation actually finishes.

The other thing I finally had to fix was to make sure that the kickstart code actually waits for DMA to be done before doing things that depend on it.  Since it's no longer part of the CPU itself (and runs in parallel to some extent), this was really needed for correctness with my new DMAgic design.